+++ src/Components/Vaccine-Administration/VaccineCalc.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import firebase from "../firebase";
+import firebase from "./firebase";
 import "./Note.css";
